Switched Cubiquity to use new 16-bit index capabilities of PolyVox.

+++ Core/CubiquityC.h
@@ -157,13 +157,13 @@ extern "C"
 	CUBIQUITYC_API int32_t cuGetNoOfIndices(uint32_t nodeHandle, uint32_t* result);
 
 	CUBIQUITYC_API int32_t cuGetVertices(uint32_t nodeHandle, float** result);
-	CUBIQUITYC_API int32_t cuGetIndices(uint32_t nodeHandle, uint32_t** result);
+	CUBIQUITYC_API int32_t cuGetIndices(uint32_t nodeHandle, uint16_t** result);
 
 	CUBIQUITYC_API int32_t cuGetNoOfVerticesMC(uint32_t nodeHandle, uint32_t* result);
 	CUBIQUITYC_API int32_t cuGetNoOfIndicesMC(uint32_t nodeHandle, uint32_t* result);
 
 	CUBIQUITYC_API int32_t cuGetVerticesMC(uint32_t nodeHandle, float** result);
-	CUBIQUITYC_API int32_t cuGetIndicesMC(uint32_t nodeHandle, uint32_t** result);
+	CUBIQUITYC_API int32_t cuGetIndicesMC(uint32_t nodeHandle, uint16_t** result);
 
 	// Clock functions
 	CUBIQUITYC_API int32_t cuGetCurrentTime(uint32_t* result);
